Refactored to add classes user type services
