+++ src/main/java/edu/umgc/librarymanager/service/LibrarianServices.java
@@ -1,7 +1,7 @@
 /*
- * Filename: ControlHelper.java
+ * Filename: LibrarianServices.java
  * Author: Scott
- * Date Created: 11/28/2020
+ * Date Created: 12/2/2020
  */
 
 package edu.umgc.librarymanager.service;
@@ -13,87 +13,19 @@ import edu.umgc.librarymanager.data.model.user.BaseUser;
 import edu.umgc.librarymanager.data.model.user.UserType;
 import edu.umgc.librarymanager.gui.DialogUtil;
 import edu.umgc.librarymanager.gui.GUIController;
-import edu.umgc.librarymanager.gui.InactivityListener;
 import edu.umgc.librarymanager.gui.MainFrame;
 import edu.umgc.librarymanager.gui.panels.PanelComposite;
-import java.util.HashMap;
 import java.util.List;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 /**
- * This class is used to provide helper methods for the GUIController class.
+ * This class is used to provide methods related to the Librarian services.
  * @author Scott
  */
-public final class ControlHelper {
+public final class LibrarianServices {
 
-    private static final Logger LOG = LogManager.getLogger(GUIController.class);
+    private LibrarianServices() {}
 
-    private ControlHelper() {}
-
-    /**
-     * This method is called when the login button is pressed.
-     * @param control The GUIController of the application.
-     */
-    public static void login(GUIController control) {
-        String username = control.getFrame().getPanelComp().getLoginPanel().getUsernameField().getText();
-        HashMap<String, BaseUser> map = control.getLogins();
-        BaseUser user = null;
-        if (map.containsKey(username)) {
-            if (map.get(username).getLogin().checkPassword(control.getFrame().getPanelComp().getLoginPanel()
-                    .getPasswordField().getPassword())) {
-                control.getFrame().getPanelComp().getLoginPanel().clearFields();
-                user = map.get(username);
-            } else {
-                DialogUtil.warningMessage("Incorrect password. Access denied.", "Login Failure");
-            }
-        } else {
-            DialogUtil.warningMessage("The username was not recognized.", "Login Failure");
-        }
-        if (user != null) {
-            control.setCurrentUser(user);
-            LOG.info(user.getUserName() + " has logged on.");
-            if (control.getCurrentUser().getUserType() == UserType.Librarian) {
-                control.getFrame().getTheMenuBar().setLibrarianMenuBar(control);
-                control.getFrame().getLayout().show(control.getFrame().getPanels(), PanelComposite.LIBRARIAN_MENU);
-            } else {
-                control.getFrame().getTheMenuBar().setPatronMenuBar(control);
-                control.getFrame().getPanelComp().getSearchPanel().setFirstName(control.getCurrentUser().getFirstName());
-                control.getFrame().getLayout().show(control.getFrame().getPanels(), PanelComposite.SEARCH);
-            }
-            control.getFrame().repaint();
-            // user is logged out after 5 min of inactivity
-            control.setInactiveListener(new InactivityListener(control.getFrame(), control.getLogoutAction(), 5));
-            control.getInactiveListener().start();
-        }
-    }
-
-    /**
-     * Performs the logout action of the user when the logout button is clicked or from inactivity.
-     * @param inactivity A boolean value for whether the logout is being done because of inactivity.
-     * @param control The GUIController for the application.
-     */
-    public static void logout(boolean inactivity, GUIController control) {
-        if (control.getCurrentUser() != null) {
-            control.getFrame().getTheMenuBar().setLoginMenuBar(control);
-            control.getFrame().getLayout().show(control.getFrame().getPanels(), PanelComposite.LOGIN);
-            control.getFrame().repaint();
-            LOG.info(control.getCurrentUser().getUserName() + " logged out.");
-            if (inactivity) {
-                LOG.info(control.getCurrentUser().getUserName() + " was logged out due to inactivity.");
-                DialogUtil.informationMessage("user was logged out due to inactivity.",
-                        "Logged Out");
-            }
-            control.setCurrentUser(null);
-            control.getInactiveListener().stop();
-            control.setInactiveListener(null);
-        }
-    }
-
-    public static void viewProfile(GUIController control) {
-        control.getFrame().getPanelComp().getUserProfilePanel().setUser(control.getCurrentUser());
-        control.getFrame().getLayout().show(control.getFrame().getPanels(), PanelComposite.PROFILE);
-    }
+    // Manage Users
 
     public static void addUser(MainFrame frame) {
         frame.getPanelComp().getAddUserPanel().setNew();
@@ -207,18 +139,7 @@ public final class ControlHelper {
         }
     }
 
-    /**
-     * Gets a hashmap of login credentials from the database.
-     * @return A HashMap with the username as the key.
-     */
-    public static HashMap<String, BaseUser> getLoginCredentials() {
-        HashMap<String, BaseUser> map = new HashMap<String, BaseUser>();
-        UserDAO userDAO = new UserDAO();
-        userDAO.openSessionwithTransaction();
-        map = userDAO.getUserHashMap();
-        userDAO.closeSessionwithTransaction();
-        return map;
-    }
+    // Manage Items
 
     /**
      * Used to manage the items of the library.
@@ -252,11 +173,6 @@ public final class ControlHelper {
             control.getFrame().getLayout().show(control.getFrame().getPanels(), PanelComposite.ALL_ITEMS);
             DialogUtil.informationMessage("The update was successful.", "Update Information");
         }*/
-    }
-
-    public static void viewSearch(GUIController control) {
-        control.getFrame().getPanelComp().getSearchPanel().reset();
-        control.getFrame().getLayout().show(control.getFrame().getPanels(), PanelComposite.SEARCH);
     }
 
 }
